pyproject.toml
[project]
name = "chroma-airlines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "fastexcel>=0.20.1",
    "polars>=1.40.1",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "transformers>=5.7.0",
]
